+++ tox.ini
@@ -149,7 +149,7 @@ commands=
     protocheck4: diff {toxinidir}/wandb/proto/v4/ {toxinidir}/wandb/proto_check/
 
 [testenv:proto{3,4}]
-basepython=python3.7
+basepython=python3
 skip_install = false
 deps =
     -r{toxinidir}/requirements.txt
